+++ commons/src/main/java/org/sentiff/gift/randomizer/commons/db/model/GiftIdea.java
@@ -1,4 +1,0 @@
-package org.sentiff.gift.randomizer.commons.db.model;
-
-public record GiftIdea(String rawText) {
-}