+++ commons/src/main/java/org/sentiff/gift/randomizer/commons/model/GiftIdea.java
@@ -1,0 +1,4 @@
+package org.sentiff.gift.randomizer.commons.model;
+
+public record GiftIdea(String rawText) {
+}